00_MLOps/04_usecases/pyproject.toml
[project]
name = "mlops-pipelines"
version = "0.1.0"
description = "MLOps Prefect Pipelines"
requires-python = ">=3.11"
dependencies = [
    "prefect>=2.14.0",
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "boto3>=1.28.0",
    "pyarrow>=14.0.0",
    "great-expectations>=0.18.0",
    "feast>=0.35.0",
    "kfp>=2.4.0",
    "mlflow>=2.9.0",
    "kubernetes>=28.0.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0.0",
]
